Rename the AHI gridded reader from AHI_grid to AHI_gridded
