+++ satpy/readers/ahi_gridded.py
